Fix all normal compile warngins Bench: 8445024

+++ src/psqt_state.hpp
@@ -32,12 +32,12 @@ public:
             auto& squares                         = pos.piece_list_sq(c);
 
             for (size_t i = 0; i < 16; i++) {
-                PieceType pt = pieces[i];
+                PieceType pt = pieces[static_cast<u8>(i)];
                 if (pt == PieceType::None) {
                     continue;
                 }
 
-                add_piece(c, pt, squares[i]);
+                add_piece(c, pt, squares[static_cast<u8>(i)]);
             }
         }
     }
@@ -86,8 +86,8 @@ private:
 
         while (valid_pieces) {
             int       i  = std::countr_zero(valid_pieces);
-            PieceType pt = pieces[i];
-            add_piece(c, pt, squares[i]);
+            PieceType pt = pieces[static_cast<u8>(i)];
+            add_piece(c, pt, squares[static_cast<u8>(i)]);
             valid_pieces &= valid_pieces - 1;
         }
     }
