+++ src/tm.cpp
@@ -15,9 +15,9 @@ time::TimePoint compute_hard_limit(time::TimePoint               search_start,
     if (settings.w_time >= 0) {
         const auto compute_buffer_time = [&]() -> u64 {
             if (stm == Color::White) {
-                return settings.w_time / 20 + settings.w_inc / 2;
+                return static_cast<u64>(settings.w_time / 20 + settings.w_inc / 2);
             } else {
-                return settings.b_time / 20 + settings.b_inc / 2;
+                return static_cast<u64>(settings.b_time / 20 + settings.b_inc / 2);
             }
         };
         hard_limit = min(hard_limit, search_start + Milliseconds(compute_buffer_time()));
